update & refactoring: checkError 모듈화 및 refactoring 진행
