+++ parser.js
@@ -1,4 +1,6 @@
 const getTokenizer = require('./tokenizer.js').getTokenizer;
+// const Square = require('./square.js');
+const checkDataError = require('./error.js').CheckError;
 
 const dataType = {
   array: 'Array Type',
@@ -14,12 +16,6 @@ const booleanType = {
   true: 'Boolean True',
   false: 'Boolean False'
 }
-
-const ERROR_MSG = {
-  BLOCK_ERROR: 'BLOCK ERROR',
-  TYPE_ERROR: '알 수 없는 타입입니다.',
-  COMMA_ERROR: '올바른 문자열이 아닙니다.'
-};
 
 class Stack {
   constructor() {
@@ -41,47 +37,6 @@ class Stack {
   }
 }
 
-class CheckError {
-  checkBlockError(arrWord) {
-    let bracketPoint = 0;
-    const splitWord = arrWord.split('');
-    const matchOpenCase = ['['];
-    const matchCloseCase = [']'];
-
-    splitWord.forEach(matchCase => {
-      if (matchOpenCase.indexOf(matchCase) > -1) bracketPoint++;
-      else if (matchCloseCase.indexOf(matchCase) > -1) {
-        if (bracketPoint === 0) throw new Error(ERROR_MSG.BLOCK_ERROR);
-        bracketPoint--;
-      }
-    });
-    if (bracketPoint === 0) return true;
-    throw new Error(ERROR_MSG.BLOCK_ERROR);
-  }
-
-  checkNumberError(temp) {
-    if (temp.match(/[0-9]\D|\D[0-9]/)) throw new Error(ERROR_MSG.TYPE_ERROR + "\nERROR_VALUE: " + temp);
-  }
-
-  checkCommaError(value) {
-    if (value.match(/['"]/m)) {
-      let commaPoint = 0;
-      const splitToken = value.split('');
-      const matchCommaCase = ['"', "'"];
-
-      splitToken.forEach(matchCase => {
-        if (matchCommaCase.indexOf(matchCase) > -1) commaPoint++;
-        else if (matchCommaCase.indexOf(matchCase) > -1) commaPoint--;
-      });
-      if (commaPoint % 2 !== 0) throw new Error(ERROR_MSG.COMMA_ERROR + "\nERROR_VALUE: " + value);
-    }
-  }
-
-  checkObjKeyError(temp) {
-    if (/['".&^%$#@!*()]/m.test(temp)) throw new Error(ERROR_MSG.TYPE_ERROR + "\nERROR_VALUE: " + temp);
-  }
-}
-
 class DataStructure {
   constructor(type, value, key) {
     this.type = type;
@@ -93,7 +48,7 @@ class DataStructure {
 
 class CheckDataType {
   constructor() {
-    this.error = new CheckError();
+    this.error = new checkDataError();
   }
   getDataType(value, stack) {
     if (this.isObjKeyValueType(value)) return this.getObjKeyValType(value, stack);
@@ -190,7 +145,7 @@ class Parser {
   }
 
   parsingObj(strData) {
-    const error = new CheckError();
+    const error = new checkDataError();
     const isError = error.checkBlockError(strData);
 
     if (isError) {
@@ -244,5 +199,5 @@ const errorcase9 = '["1a"a"a"s""3",[22,23,[11,[112233],112],55],33]';
 // const errorTest9 = parsingObj(errorcase9); // COMMA ERROR => "1a"a"a"s""3"
 
 const parser = new Parser();
-const result = parser.parsingObj(testcase8)
+const result = parser.parsingObj(testcase9)
 console.log(JSON.stringify(result, null, 2));
